refactor

+++ commands/admin/clear.js
@@ -2,19 +2,22 @@
 
 module.exports = {
   name: 'clear',
-  description: 'Rensar alla meddelande som har med botten att göra.',
+  description: 'Rensar alla meddelande till och från boten.',
   aliases: [ 'rensa' ],
   group: 'admin',
-  usage: 'clear',
+  usage: 'clear <all>',
+  example: 'clear',
   serverOnly: true,
   adminOnly: true,
   async execute( message, args ) {
-    message.delete();
+    if ( !message.deleted ) message.delete();
     let messages = await message.channel.fetchMessages( { limit: 100 } );
-    messages = messages.filter( function ( m ) {
-      if ( m.author.bot ) return true;
-      else if ( config.prefix.indexOf( m.content.charAt( 0 ) ) !== -1 ) return true;
-    } );
+    if ( args[ 0 ] !== 'all' ) {
+      messages = messages.filter( function ( m ) {
+        if ( m.author.bot ) return true;
+        else if ( config.prefix.indexOf( m.content.charAt( 0 ) ) !== -1 ) return true;
+      } );
+    }
     message.channel.bulkDelete( messages );
   },
 };
